#3 天気予報を扱うクラスの名前が誤っていたので修正
