--- src/components/organisms/WeekdayTemperatureForecast.stories.tsx
@@ -1,9 +1,0 @@
-import { FC } from 'react';
-import WeekdayTemperatureForecast from './WeekdayTemperatureForecast';
-
-export default {
-  component: WeekdayTemperatureForecast,
-  title: 'WeekdayTemperatureForecast',
-};
-
-export const Defualt: FC = () => <WeekdayTemperatureForecast />;
