+++ src/components/organisms/WeekdayWeatherForecast.stories.tsx
@@ -1,0 +1,9 @@
+import { FC } from 'react';
+import WeekdayWeatherForecast from './WeekdayWeatherForecast';
+
+export default {
+  component: WeekdayWeatherForecast,
+  title: 'WeekdayTemperatureForecast',
+};
+
+export const Defualt: FC = () => <WeekdayWeatherForecast />;
